+++ src/pages/FriendlyCampus/CampusLaw.js
@@ -8,7 +8,7 @@ class CampusLaw extends Component {
     static navigationOptions = ({ navigation, screenProps }) => ({
         title: '校園法律宣導',
         headerRight: <Icon containerStyle={{ marginRight: 5 }} name='home' onPress={() => navigation.dispatch(
-            NavigationActions.reset({index: 0, actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
+            NavigationActions.back({index: 0, actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
         )} />
     });