+++ src/pages/FriendlyCampus/GenderEquality.js
@@ -8,7 +8,7 @@ class GenderEquality extends Component {
     static navigationOptions = ({ navigation, screenProps }) => ({
         title: '性別平等教育',
         headerRight: <Icon containerStyle={{ marginRight: 5 }} name='home' onPress={() => navigation.dispatch(
-            NavigationActions.reset({index: 0, actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
+            NavigationActions.back({index: 0, actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
         )} />
     });