+++ src/pages/FriendlyCampus/StudentsAppeals.js
@@ -8,7 +8,7 @@ class StudentsAppeals extends Component {
     static navigationOptions = ({ navigation, screenProps }) => ({
         title: '學生申訴',
         headerRight: <Icon containerStyle={{ marginRight: 5 }} name='home' onPress={() => navigation.dispatch(
-            NavigationActions.reset({index: 0, actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
+            NavigationActions.back({index: 0, actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
         )} />
     });