+++ src/pages/LearningDevelopment/Course/GeneralCourse.js
@@ -9,7 +9,7 @@ class GeneralCourse extends Component {
     static navigationOptions = ({ navigation, screenProps }) => ({
         title: '校定畢業門檻',
         headerRight: <Icon containerStyle={{ marginRight: 5 }} name='home' onPress={() => navigation.dispatch(
-            NavigationActions.reset({index: 0, actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
+            NavigationActions.back({index: 0, actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
         )} />
     });