+++ src/pages/LearningDevelopment/LibraryResources.js
@@ -8,7 +8,7 @@ class LibraryResources extends Component {
     static navigationOptions = ({ navigation, screenProps }) => ({
         title: '圖書資源',
         headerRight: <Icon containerStyle={{ marginRight: 5 }} name='home' onPress={() => navigation.dispatch(
-            NavigationActions.reset({index: 0, actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
+            NavigationActions.back({index: 0, actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
         )} />
     });