+++ src/pages/LearningDevelopment/Takeoff.js
@@ -9,7 +9,7 @@ class Takeoff extends Component {
     static navigationOptions = ({ navigation, screenProps }) => ({
         title: '學生請假',
         headerRight: <Icon containerStyle={{ marginRight: 5 }} name='home' onPress={() => navigation.dispatch(
-            NavigationActions.reset({index: 0, actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
+            NavigationActions.back({index: 0, actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
         )} />
     });