+++ src/pages/MainItem/LearningDevelopment.js
@@ -41,7 +41,7 @@ class LearningDevelopment extends Component {
     static navigationOptions = ({ navigation, screenProps }) => ({
         title: '學習發展',
         headerRight: <Icon containerStyle={{ marginRight: 5 }} name='home' onPress={() => navigation.dispatch(
-            NavigationActions.reset({index: 0, actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
+            NavigationActions.back({index: 0, actions: [NavigationActions.navigate({routeName: 'Homepage'})]})
         )} />
     });